requirements.txt
google-generativeai
python-dotenv
duckduckgo-search
colorama
streamlit